fastapi_ws_yolo/app.py
import asyncio
import contextlib
from pathlib import Path
from fastapi import FastAPI, File, UploadFile, WebSocket, WebSocketDisconnect
from PIL import Image
from fastapi.responses import FileResponse
from fastapi.staticfiles import StaticFiles
from object_detection_model import ObjectDetection, Objects, receive, detect


object_detection = ObjectDetection()


@contextlib.asynccontextmanager
async def lifespan(app: FastAPI):
    global object_detection
    object_detection.load_model()
    yield


app = FastAPI(lifespan=lifespan)

@app.post("/object-detection-static", response_model=Objects)
async def post_object_detection(image: UploadFile = File(...)) -> Objects:
    image_object = Image.open(image.file)
    return object_detection.predict(image_object)



@app.websocket("/object-detection")
async def ws_object_detection(websocket: WebSocket):
    global object_detection
    await websocket.accept()
    queue: asyncio.Queue = asyncio.Queue(maxsize=1)
    receive_task = asyncio.create_task(receive(websocket, queue))
    detect_task = asyncio.create_task(detect(websocket, queue, object_detection))
    try:
        done, pending = await asyncio.wait(
            {receive_task, detect_task},
            return_when=asyncio.FIRST_COMPLETED,
        )
        for task in pending:
            task.cancel()
        for task in done:
            task.result()
    except WebSocketDisconnect:
        pass


@app.get("/")
async def index():
    return FileResponse(Path(__file__).parent / "index.html")


static_files_app = StaticFiles(directory=Path(__file__).parent / "assets")
app.mount("/assets", static_files_app)


