fastapi_ws_yolo/requirements.txt
python-decouple
torch --index-url https://download.pytorch.org/whl/cpulangchain
fastapi
uvicorn[standard]
transformers[torch]
Pillow
python-multipart